Refactoring NamingConventionRules
